requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.5.4 # For Gemini